fix: constant pool u16, var re-decl, smoke test
- Change constant pool index from u8 to u16 (max 65536 constants)
- Update compiler to emit u16 for all constant indices
- Update VM to read u16 for all constant indices
- Fix var re-declaration in functions to reuse existing slot
- Add vm_smoke_test.js (55 assertions covering all VM features)

+++ src/core/opcode.rs
@@ -113,14 +113,14 @@ impl<'gc> Chunk<'gc> {
     pub fn write_opcode(&mut self, opcode: Opcode) {
         self.write_byte(opcode as u8);
     }
-    
+
     pub fn write_u16(&mut self, value: u16) {
         self.code.push((value & 0xff) as u8);
         self.code.push(((value >> 8) & 0xff) as u8);
     }
-    
-    pub fn add_constant(&mut self, value: Value<'gc>) -> u8 {
+
+    pub fn add_constant(&mut self, value: Value<'gc>) -> u16 {
         self.constants.push(value);
-        (self.constants.len() - 1) as u8
+        (self.constants.len() - 1) as u16
     }
 }
